Move parse code to Choice
Choice can be used in multiple palces, e.g. element group and complex
type. It's therefore better to have code there for reuse.
